test: make wide-depth test data generators 32-bit safe
Address CodeRabbit review on PR #5:
- golden_baseline_test.go: compute the PCM generator in uint32 so the golden
  hashes are stable on 32-bit (GOARCH 386/arm) as well as 64-bit. The extracted
  byte is unchanged on 64-bit (it reads bits below 2^32), so TestInt32OutputUnchanged
  still matches the captured baseline.
- analyze64_test.go: the seed multiplier exceeds 2^31, which overflows a 32-bit
  int constant and fails to compile on 386/arm; compute it in uint32.

Verified: golden hashes unchanged, parity test passes, both packages compile
under GOARCH=386.

+++ internal/lpc/analyze64_test.go
@@ -13,7 +13,9 @@ func TestAnalyzeLPCGenericParity(t *testing.T) {
 	s32 := make([]int32, n)
 	s64 := make([]int64, n)
 	for i := range n {
-		v := int32((i*2654435761)>>20) & 0x3FFFFF // ~22-bit, fits both
+		// uint32 arithmetic avoids overflowing a 32-bit int (the multiplier exceeds
+		// 2^31), so the test compiles and runs on 32-bit targets too.
+		v := int32((uint32(i)*2654435761)>>20) & 0x3FFFFF // ~22-bit, fits both
 		s32[i] = v
 		s64[i] = int64(v)
 	}
